+++ packages/ui/src/components/Table/TableRow.tsx
@@ -1,9 +1,10 @@
-import { ReactNode } from 'react';
+import { ReactNode, AriaRole } from 'react';
 import styled from '@emotion/styled';
 
 import { Stack } from '../Stack';
 
 export type TableRowProps = {
+  header?: boolean;
   columns: ReactNode[];
 };
 
@@ -16,11 +17,23 @@ const Row = styled(Stack)(({ theme }) => ({
   padding: theme.spacing(0, 2),
 }));
 
-export const TableRow = ({ columns }: TableRowProps) => (
-  <Row direction="row" align="center">
-    <Column style={{ width: 90 }}>{columns[0]}</Column>
-    <Column style={{ flex: 1 }}>{columns[1]}</Column>
-    <Column style={{ justifyContent: 'end' }}>{columns[2]}</Column>
-    <Column style={{ width: 90, justifyContent: 'end' }}>{columns[3]}</Column>
-  </Row>
-);
+export const TableRow = ({ columns, header = false }: TableRowProps) => {
+  const columnRole: AriaRole = header ? 'columnheader' : 'cell';
+
+  return (
+    <Row role="row" direction="row" align="center">
+      <Column role={columnRole} style={{ width: 90 }}>
+        {columns[0]}
+      </Column>
+      <Column role={columnRole} style={{ flex: 1 }}>
+        {columns[1]}
+      </Column>
+      <Column role={columnRole} style={{ justifyContent: 'end' }}>
+        {columns[2]}
+      </Column>
+      <Column role={columnRole} style={{ width: 90, justifyContent: 'end' }}>
+        {columns[3]}
+      </Column>
+    </Row>
+  );
+};