+++ packages/ui/src/createContext.ts
@@ -30,10 +30,17 @@ export type ReportingContext = {
 };
 
 export type GameInfoContext = {
+  loading: boolean;
   name?: string;
   url?: string;
   logoUrl?: string;
   tags?: string[];
+
+  preloader: {
+    url?: string;
+    title?: string;
+    description?: string;
+  };
 };
 
 export type ContextState = {
@@ -76,8 +83,10 @@ const defaultState: ContextState = {
     isReady: false,
   },
   gameInfo: {
+    loading: false,
     name: '',
     tags: [''],
+    preloader: {},
   },
 };