+++ packages/ui/src/icons/TrophyIcon.tsx
@@ -1,6 +1,6 @@
 import { SVGProps } from 'react';
 
-import trophy from '../assets/trophy.png';
+import { trophy } from '../assets';
 
 export const TrophyIcon = (props: SVGProps<SVGSVGElement>) => (
   <svg xmlns="http://www.w3.org/2000/svg" width="1em" height="1em" viewBox="20 20 160 160" {...props}>