+++ tests/wdio/objects/ConnectWalletModal.ts
@@ -5,7 +5,25 @@ export class ConnectWalletModal extends Widget {
     return $('cere-fullscreen-modal cere-connect-wallet');
   }
 
-  get connectButton() {
-    return this.shadowRoot.findByTestId$('connectWallet');
+  get claimButton() {
+    return this.shadowRoot.findByRole$('button', { name: /Claim \d+ free tokens/i });
+  }
+
+  get score() {
+    return this.shadowRoot.findByRole$('heading', { name: /Your score: \d+/i });
+  }
+
+  async getScore() {
+    const text = await this.score.getText();
+    const [match, amount] = text.match(/(\d+)/i) || [];
+
+    return match ? +amount : undefined;
+  }
+
+  async getRewardAmount() {
+    const text = await this.claimButton.getText();
+    const [match, amount] = text.match(/(\d+)/i) || [];
+
+    return match ? +amount : undefined;
   }
 }