+++ tests/wdio/objects/LeaderboardModal.ts
@@ -6,7 +6,7 @@ export class LeaderboardModal extends Widget {
   }
 
   get walletAddress() {
-    return this.shadowRoot.findByLabelText$('Wallet address');
+    return this.shadowRoot.findByLabelText$(/wallet address/i);
   }
 
   get table() {
@@ -17,7 +17,46 @@ export class LeaderboardModal extends Widget {
     return this.table.findByRole$('row', { selected: true });
   }
 
+  get tokensBalance() {
+    return this.shadowRoot.findByText$(/tokens balance: (\d+\.\d{2})/i);
+  }
+
   get playAgainButton() {
-    return this.shadowRoot.findByRole$('button', { name: 'Play again' });
+    return this.shadowRoot.findByRole$('button', { name: /play again/i });
+  }
+
+  async getRewardNotificationAmount() {
+    const notification = await browser.waitUntil(() => this.shadowRoot.queryByRole('alert'), {
+      timeout: 30000,
+      timeoutMsg: 'The reward notification did not appear in time',
+    });
+
+    const text = await notification.getText();
+    const [match, amount] = text.match(/(\d+) \$CERE tokens/i) || [];
+
+    return match ? +amount : undefined;
+  }
+
+  async getWalletAddress() {
+    const element = await this.walletAddress;
+
+    return element.getAttribute('title');
+  }
+
+  async getBalance() {
+    const text = await this.tokensBalance.getText();
+    const [match, amount] = text.match(/(\d+\.\d{2})/i) || [];
+
+    return match ? +amount : undefined;
+  }
+
+  async getActiveRowData() {
+    const [rank, player, , score] = await this.activeRow.getAllByRole$('cell');
+
+    return {
+      score: await score.getText().then(Number),
+      rank: await rank.getText().then(Number),
+      address: await player.getByLabelText$(/address/i).getAttribute('data-full'),
+    };
   }
 }