+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 ### vNext
 
-- Add WebDriverIO auto test
+-
+
+### v0.4.1
+
+- Fix preloader data loading state
+- Configure auto-tests to work in VPN protected environment
+
+### v0.4.0
+
+- [[Feature](https://www.notion.so/cere/Cere-Wallet-Achieve-99-new-users-login-success-rate-in-Cere-Wallet-20610fe9e8564934b544a28aee4dacb9?pvs=4)] Update Cere Wallet SDK to switch to web3auth Cyan network
+- [[Feature]](https://www.notion.so/cere/UI-tests-for-Wallet-Client-and-Games-49a07b7b50114f49bd4c012c6e6220f5?pvs=4) Add WebDriverIO auto test
 - Add the SDK integration run-able examples (`/examples`)
+- [[Feature]](https://www.notion.so/cere/Update-Leaderboard-modal-to-show-tournaments-65b3a7f4d8d848a384d36618e01b338c?pvs=4) Update Leaderboard modal to show tournaments
 
 ### v0.3.0