+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Release Notes:
 
+### v0.4.1
+
+- Fix preloader data loading state
+
+### v0.4.0
+
+- Update Cere Wallet SDK
+- Update Leaderboard modal to show tournaments
+
 ### v0.3.0
 
 - Fix ran icon issue on Leaderboard